uspace: achieve a zero-warning build and enforce it with -Werror

Resolve all -Wall -Wextra warnings across gcc and clang (xdp and no-xdp)
and turn on -Werror in the build-uspace CI job so new warnings fail the
build instead of accumulating.

- Fix compiler warnings toward a zero-warning build, including a real
  bug where ec_eoe_init() passed a computed name as a printf format
  string (now "%s", name); NUL-safe copies, sign-compare casts, and
  full struct initializers.
- Promote warnings to errors in CI, keeping only -Wno-unused-parameter
  (uniform FSM state-fn signatures) and -Wno-type-limits (deliberate
  EC_*_DBG tautology) suppressed.
- Check chdir/freopen return values in daemonize() to satisfy
  -Wunused-result under Ubuntu gcc's _FORTIFY_SOURCE.
- Bound the interface copy into IFNAMSIZ ifr_name with "%.*s" after the
  uspace merge widened ec_transport.interface, silencing gcc-13
  -Wformat-truncation.

Refs #182

+++ master/fsm_foe.c
@@ -231,8 +231,8 @@ int ec_foe_prepare_data_send(
 
     remaining_size = fsm->tx_buffer_size - fsm->tx_buffer_offset;
 
-    if (remaining_size < fsm->slave->configured_tx_mailbox_size
-            - EC_MBOX_HEADER_SIZE - EC_FOE_HEADER_SIZE) {
+    if (remaining_size < (size_t) (fsm->slave->configured_tx_mailbox_size
+            - EC_MBOX_HEADER_SIZE - EC_FOE_HEADER_SIZE)) {
         current_size = remaining_size;
         fsm->tx_last_packet = 1;
     } else {
